feat: move example color palette to ColorPicker story
