+++ src/components/ColorInputFlyout/ColorPickerFlyout.stories.tsx
@@ -1,9 +1,9 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 import { Meta, Story } from "@storybook/react";
-import { EXAMPLE_PALETTES } from "@utilities/colors";
 import React, { useState } from "react";
 import { Color } from "../../types/colors";
+import { EXAMPLE_PALETTES } from "../ColorPicker/ColorPicker.stories";
 import { ColorPickerFlyout as ColorPickerFlyoutComponent, ColorPickerFlyoutProps } from "./ColorPickerFlyout";
 
 // eslint-disable-next-line import/no-default-export
