+++ src/components/ColorPicker/ColorPicker.stories.tsx
@@ -1,10 +1,46 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 import { Meta, Story } from "@storybook/react";
-import { EXAMPLE_PALETTES } from "@utilities/colors";
 import React, { useState } from "react";
-import { ColorFormat } from "../../types/colors";
+import tinycolor from "tinycolor2";
+import { Color, ColorFormat, Palette } from "../../types/colors";
 import { ColorPicker, ColorPickerProps } from "./ColorPicker";
+
+const generatePalette = (color: string, amount: number): Color[] => {
+    const sourceColor = tinycolor(color);
+    const palette = [...new Array(amount)].map((_, index) => {
+        const name = (90 - index * 10).toString();
+        const lightColor = sourceColor.lighten(index * 3).toRgb();
+        return {
+            ...lightColor,
+            name,
+        };
+    });
+    return palette;
+};
+
+export const EXAMPLE_PALETTES: Palette[] = [
+    {
+        id: "red",
+        title: "Red",
+        source: "#992136",
+    },
+    {
+        id: "green",
+        title: "Green",
+        source: "#006452",
+    },
+    {
+        id: "yellow",
+        title: "Yellow",
+        source: "#cc9000",
+    },
+].map((palette) => {
+    return {
+        ...palette,
+        colors: generatePalette(palette.source, 6),
+    };
+});
 
 // eslint-disable-next-line import/no-default-export
 export default {
